requirements.txt
--extra-index-url https://download.pytorch.org/whl/cu118
torch
torchaudio
selenium
datasets
sentencepiece
soundfile
botocore==1.34.88
python-telegram-bot
twilio
flask
quart